Avoid dereferencing nullptr C-style strings
Added runtime checks for C-style strings (char*) being nullptr during matching and serialization.

This fix prevents nullptr dereferences in the case that the null character pointer (as opposed to the null nullptr_t) is expected and in the case that a non-null string is expected but nullptr is actually passed.